# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /churches/{churchId}/integrated-attempts:
    post:
      operationId: startIntegratedAttempt
      tags: [profiles]
      summary: Start a frozen anonymous adult pilot assessment
      parameters:
        - { name: churchId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntegratedAttemptStartInput" }
      responses:
        "201":
          description: Draft created; token is returned only once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegratedAttemptView" }
        "400": { description: Adult confirmation or church configuration invalid }
        "404": { description: Pilot not enabled }
        "429": { description: Too many drafts }
  /churches/{churchId}/integrated-attempts/{attemptId}:
    parameters:
      - { name: churchId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      - { name: attemptId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: Authorization, in: header, required: true, schema: { type: string, pattern: "^Bearer [A-Za-z0-9_-]{43}$" } }
    get:
      operationId: getIntegratedAttempt
      tags: [profiles]
      summary: Resume the original frozen adult assessment
      responses:
        "200":
          description: Saved draft without credential hash or scoring maps
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegratedAttemptView" }
        "404": { description: Draft not found or unauthorized }
        "410": { description: Draft expired }
        "429": { description: Request rate exceeded }
    patch:
      operationId: saveIntegratedAttempt
      tags: [profiles]
      summary: Atomically save bounded progress at the expected revision
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntegratedAttemptSaveInput" }
      responses:
        "200":
          description: Progress saved with incremented revision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegratedAttemptView" }
        "400": { description: Invalid answer or form progress }
        "404": { description: Draft not found or unauthorized }
        "409": { description: Revision conflict or already completed; saved answers unchanged }
        "410": { description: Draft expired }
        "429": { description: Request rate exceeded }
...
components:
  schemas:
...
    IntegratedAttemptStartInput:
      type: object
      additionalProperties: false
      required: [adultConfirmed, celibacyEligible, optionalExperienceOptIn]
      properties:
        adultConfirmed: { type: boolean, enum: [true] }
        celibacyEligible: { type: boolean }
        optionalExperienceOptIn: { type: boolean }
    IntegratedAnswers:
      type: object
      additionalProperties:
        oneOf:
          - { type: integer, minimum: 1, maximum: 5 }
          - { type: string, enum: [na, skip] }
    IntegratedAttemptSaveInput:
      type: object
      additionalProperties: false
      required: [revision, answers, formState]
      properties:
        revision: { type: integer, minimum: 0, maximum: 2147483646 }
        answers: { $ref: "#/components/schemas/IntegratedAnswers" }
        formState: { type: object, additionalProperties: true, description: "Allowlisted non-assessment wizard fields, at most 48KB." }
    IntegratedQuestionView:
      type: object
      required: [id, text, responseModel]
      properties:
        id: { type: string }
        text: { type: string }
        responseModel: { type: string, enum: [reflectionLikert, personalityBipolar, specialExperienceLikert] }
        poles: { type: array, items: { type: string }, minItems: 2, maxItems: 2 }
    IntegratedAttemptView:
      type: object
      required: [attemptId, revision, expiresAt, status, bankVersion, scoringVersion, snapshot, questions, answers, formState, responseModels]
      properties:
        attemptId: { type: string, format: uuid }
        token: { type: string, description: "Present only at creation. Treat as a private bearer credential." }
        revision: { type: integer }
        expiresAt: { type: string, format: date-time }
        status: { type: string, enum: [draft, completed] }
        profileId: { type: ["integer", "null"] }
        bankVersion: { type: string }
        scoringVersion: { type: string }
        snapshot: { type: object, additionalProperties: true }
        responseModels: { type: object, additionalProperties: true }
        questions: { type: array, items: { $ref: "#/components/schemas/IntegratedQuestionView" } }
        answers: { $ref: "#/components/schemas/IntegratedAnswers" }
        formState: { type: object, additionalProperties: true }
...