# lib/api-zod/src/generated/types/integratedAttemptStartInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IntegratedAttemptStartInput {
  adultConfirmed: true;
  celibacyEligible: boolean;
  optionalExperienceOptIn: boolean;
}