# lib/api-zod/src/generated/types/integratedQuestionView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IntegratedQuestionViewResponseModel } from './integratedQuestionViewResponseModel';

export interface IntegratedQuestionView {
  id: string;
  text: string;
  responseModel: IntegratedQuestionViewResponseModel;
  /**
     * @minItems 2
     * @maxItems 2
     */
  poles?: string[];
}